Moved estimateMu() to main area
Started integration of new code in the core/ folder toward the main
R/ folder: replaced each occurrence of the mu_est() method with the
new estimateMu().

Usage of the new method of the COTAN class in the old scCOTAN class
required possibility of conversion. Since the two classes represent
essentially the same object, it has been decided that it would be
smoother if implicit casting of a scCOTAN object to the COTAN class
was allowed. This entailed definition of an appropariate setIs()
method