Moved nu, lambda, dispersion into meta
+ Moved the slots nu, lambda and dispersion to be columns in the
corresponding metaCells and metaGenes data.frames.

+ Realigned the COTAN accessors to the new structure.

+ Added more stringent checks regarding clusterizations in
metaCells: now those are recognised by the fact that their name
starts with 'CL_'; also now for each such columns an entry in the
clustersCoex must be present (even if empty).

+ Renamed addColumnToDF() to setColumnInDF() as it is now also
able to reset the stored values and not just add a new column