Implemented coex synchronization flags
Implemented mechanism by which the COTAN object can keep track of
whether the coex slots are still in line with the underlying
parameters (lambda, dispersion, nu).

If any of these parameters are changed after any of the 'coex'
matrices have been calculated, the matrices are flagged as
out-of-sync and thus unreachable via 'get' methods.