Refactored cellsUniformClustering()
Renamed and made a free function the method called:
cell_homogenoeus_clustering().

Refactored, renamed and made into a free function also the method:
cluster_homogenneity_check().
The new name is:
checkCluesterUniformity()

Rationalised the messages in the functions in order not to be too
overwelming in normal usage.

Used roman numerals for the final clusters.