Made `torch` support as opt-in
Now COTAN will use try to use the `torch` library only if
the option "COTAN.UseTorch" is set to `TRUE`.

This is necessary since, as of now, it is not possible to check
whether the torch installation actually ready to use without
hard failures. This might change in the future and have the feature
as opt-out instead.